Have refactored AKPlayer into manageable extensions
