+++ AudioKit/Common/Nodes/Playback/Players/Player/AKPlayer+Timing.swift
@@ -1,0 +1,43 @@
+//
+//  AKPlayer+Timing.swift
+//  AudioKit
+//
+//  Created by Ryan Francesconi on 6/12/18.
+//  Copyright © 2018 AudioKit. All rights reserved.
+//
+
+extension AKPlayer: AKTiming {
+    public func start(at audioTime: AVAudioTime?) {
+        play(at: audioTime)
+    }
+
+    public var isStarted: Bool {
+        return isPlaying
+    }
+
+    public func setPosition(_ position: Double) {
+        startTime = position
+        if isPlaying {
+            stop()
+            play()
+        }
+    }
+
+    public func position(at audioTime: AVAudioTime?) -> Double {
+        guard let playerTime = playerNode.playerTime(forNodeTime: audioTime ?? AVAudioTime.now()) else {
+            return startTime
+        }
+        return startTime + Double(playerTime.sampleTime) / playerTime.sampleRate
+    }
+
+    public func audioTime(at position: Double) -> AVAudioTime? {
+        let sampleRate = playerNode.outputFormat(forBus: 0).sampleRate
+        let sampleTime = (position - startTime) * sampleRate
+        let playerTime = AVAudioTime(sampleTime: AVAudioFramePosition(sampleTime), atRate: sampleRate)
+        return playerNode.nodeTime(forPlayerTime: playerTime)
+    }
+
+    open func prepare() {
+        preroll(from: startTime, to: endTime)
+    }
+}
